Make amount int

+++ src/paper/related_models/apc_model.py
@@ -4,7 +4,7 @@ from django.db import models
 # Model for article processing charges (APC) for papers
 class PaperAPC(models.Model):
     paper = models.ForeignKey("Paper", on_delete=models.CASCADE, related_name="apcs")
-    amount = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
+    amount = models.IntegerField(null=True, blank=True)
     currency = models.CharField(max_length=3, null=True, blank=True)
     paid = models.BooleanField(default=False)
     created_date = models.DateTimeField(auto_now_add=True)
